Add zero file length check to make sure pesign didn't fail (rhbz 991808)

+++ kernel.spec
@@ -1609,6 +1609,10 @@ BuildKernel() {
     %if %{signmodules}
     # Sign the image if we're using EFI
     %pesign -s -i $KernelImage -o vmlinuz.signed
+    if [ ! -s vmlinuz.signed ]; then
+        echo "pesigning failed"
+        exit 1
+    fi
     mv vmlinuz.signed $KernelImage
     %endif
     $CopyKernel $KernelImage \
@@ -2241,6 +2245,9 @@ fi
 #                 ||----w |
 #                 ||     ||
 %changelog
+* Wed Aug 07 2013 Josh Boyer <jwboyer@redhat.com>
+- Add zero file length check to make sure pesign didn't fail (rhbz 991808)
+
 * Tue Aug 06 2013 Josh Boyer <jwboyer@redhat.com> - 3.11.0-0.rc4.git1.1
 - Linux v3.11-rc4-20-g0fff106
 - Reenable debugging options.
